+++ app/src/main/java/com/example/k/superbag2/bean/UploadItembean.java
@@ -23,13 +23,13 @@ public class UploadItembean extends BmobObject {
     private String feelings;
     private int importance;
     private String phone;
-    private float updateTime;
+    private long updateTime;
 
-    public float getUpdateTime() {
+    public long getUpdateTime() {
         return updateTime;
     }
 
-    public void setUpdateTime(float updateTime) {
+    public void setUpdateTime(long updateTime) {
         this.updateTime = updateTime;
     }
